+++ src/features/order/hooks/useStreamOrderStatus.ts
@@ -1,8 +1,8 @@
 import { useEffect } from 'react'
-import { OrderStatus } from '../types'
 import { useOrderStatusStore } from '@/store/orderStatus'
 import { BASE_URL } from '@/constants/api'
 import { EventSourcePolyfill } from 'event-source-polyfill'
+import { OrderStatus } from '../types'
 
 interface UseStreamOrderStatusProps {
   orderId: number
